+++ Apps.OpenAI/Actions/TranslationActions.cs
@@ -19,12 +19,10 @@ using Blackbird.Applications.Sdk.Common.Invocation;
 using Blackbird.Applications.Sdk.Glossaries.Utils.Dtos;
 using Blackbird.Applications.SDK.Blueprints;
 using Blackbird.Applications.SDK.Extensions.FileManagement.Interfaces;
-using Blackbird.Filters.Bilingual.Xliff1;
 using Blackbird.Filters.Constants;
 using Blackbird.Filters.Enums;
 using Blackbird.Filters.Extensions;
 using Blackbird.Filters.Transformations;
-using RestSharp;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -194,24 +192,7 @@ public class TranslationActions(InvocationContext invocationContext, IFileManage
         }
 
         result.TargetsUpdatedCount = updatedCount;
-
-        if (input.OutputFileHandling == "original")
-        {
-            var targetContentResult = content.Target();
-            if (!targetContentResult.Success)
-                throw new PluginMisconfigurationException(targetContentResult.Error);
-            var targetContent = targetContentResult.Value;
-            result.File = await FileManagementClient.UploadAsync(targetContent.ToStream(), targetContent.OriginalMediaType, targetContent.OriginalName);
-        }
-        else if (input.OutputFileHandling == "xliff1")
-        {
-            var xliff1String = Xliff1Serializer.Serialize(content);
-            result.File = await FileManagementClient.UploadAsync(xliff1String.ToStream(), MediaTypes.Xliff1, content.BilingualFileName);
-        }
-        else
-        {
-            result.File = await FileManagementClient.UploadAsync(content.ToStream(), MediaTypes.Xliff2, content.BilingualFileName);
-        }
+        result.File = await OutputFileHandler.ToOutputFile(FileManagementClient, content, input.OutputFileHandling);
 
         return result;
     }
